+++ src/shared/context/index.ts
@@ -1,1 +1,2 @@
 export * from './AuthContext';
+export * from './ThemeContext';
