+++ src/shared/lib/index.ts
@@ -2,9 +2,9 @@ export { queryClient } from './config/queryClient';
 export * from './book';
 export * from './number';
 export * from './record';
+export * from './style';
 export * from './validation';
 export * from './config/supabase';
-export * from './hooks/useThemeMode';
 export * from './hooks/useImageFileUploader';
 export * from './hooks/useModal';
 export * from './hooks/useOnClickOutside';
