+++ src/shared/lib/style.ts
@@ -1,0 +1,11 @@
+export const getInitialThemeMode = () => {
+  const localStorageTheme = localStorage.getItem('theme');
+  const initialThemeMode =
+    localStorageTheme === 'dark' ||
+    (!localStorageTheme &&
+      window.matchMedia('(prefers-color-scheme: dark)').matches)
+      ? 'dark'
+      : 'light';
+
+  return initialThemeMode;
+};
