+++ src/shared/ui/skeleton/SkeletonWrapper.tsx
@@ -1,14 +1,14 @@
-import { ReactNode } from 'react';
+import { ReactNode, useContext } from 'react';
 import { SkeletonTheme } from 'react-loading-skeleton';
 
-import { useThemeMode } from '@/shared/lib';
+import { ThemeContext } from '@/shared/context';
 
 interface SkeletonWrapperProps {
   children: ReactNode;
 }
 
 export const SkeletonWrapper = ({ children }: SkeletonWrapperProps) => {
-  const { themeMode } = useThemeMode();
+  const { themeMode } = useContext(ThemeContext);
 
   const baseColor = themeMode === 'dark' ? '#333' : undefined;
   const highlightColor = themeMode === 'dark' ? '#555' : undefined;
